refactor: centralize role ranks and membership list queries for DRY.
